Rewrite ArmoredPacketReader not to use ArmoredInputStream and to track the state in more organized way
It may not handle all the edge cases just yet and the performance is not great. One trick that could
be used to improve the performance is to opportunistically read as much bytes as is the size of the
footer (-----END ...-----) since we are always going to read that anyway. That would likely require
restructuring the code a bit or using System.IO.Pipelines.